+++ consensus/bor/clerk/clerk.go
@@ -1,4 +1,4 @@
-package bor
+package clerk
 
 import (
 	"fmt"
