+++ consensus/bor/errors.go
@@ -3,6 +3,8 @@ package bor
 import (
 	"fmt"
 	"time"
+
+	"github.com/ethereum/go-ethereum/consensus/bor/clerk"
 )
 
 // TotalVotingPowerExceededError is returned when the maximum allowed total voting power is exceeded
@@ -129,7 +131,7 @@ type InvalidStateReceivedError struct {
 	Number      uint64
 	LastStateID uint64
 	To          *time.Time
-	Event       *EventRecordWithTime
+	Event       *clerk.EventRecordWithTime
 }
 
 func (e *InvalidStateReceivedError) Error() string {
