+++ consensus/bor/genesis.go
@@ -1,0 +1,15 @@
+package bor
+
+import (
+	"math/big"
+
+	"github.com/ethereum/go-ethereum/consensus/bor/clerk"
+	"github.com/ethereum/go-ethereum/consensus/bor/statefull"
+	"github.com/ethereum/go-ethereum/core/state"
+	"github.com/ethereum/go-ethereum/core/types"
+)
+
+type GenesisContract interface {
+	CommitState(event *clerk.EventRecordWithTime, state *state.StateDB, header *types.Header, chCtx statefull.ChainContext) (uint64, error)
+	LastStateId(snapshotNumber uint64) (*big.Int, error)
+}
